src/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
prometheus-client
celery
pika
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-celery
# socket
# time
# sys